bug deletemessage

+++ TripleZero.Repository/Repository/Dto/ArenaDto.cs
@@ -7,7 +7,7 @@ namespace TripleZero.Repository.Dto
     public class ArenaDto
     {
         public int CurrentRank { get; set; }
-        public int AverageRank { get; set; }
+        public double AverageRank { get; set; }
         public List<string> ArenaTeam { get; set; }
     }
 }
